支持视频详情页播放列表形式显示合集视频

+++ src/BiliLite.UWP/Controls/PlayerControl.xaml.cs
@@ -110,7 +110,7 @@ namespace BiliLite.Controls
         /// <summary>
         /// 时长（毫秒）
         /// </summary>
-        public int duration { get; set; }
+        public long duration { get; set; }
         public LocalPlayInfo LocalPlayInfo { get; set; }
         public object parameter { get; set; }
